Changed some small effects around sharing/saving custom recipes
Changed "Save" button to set to "Saved" when a custom recipe is saved, and "Share" button && "Save" button to change to "Shared"+"Saved" respectively when it is shared. Share button now also saves the recipe, and lets the user know that it has done so.

Also changed some styling in the shareViaDM displayOnlineUsers to avoid multiple onClicks being excuted.